Enums now work (but not arrays of Enums). BaseType::width(bool = false) is now a const method and Array::width() is no more (it didn't do anything that Vector::width() didn't do with fewer lines of code).

+++ unit-tests/ServerFunctionsListUnitTest.cc
@@ -122,7 +122,7 @@ public:
             return;
         }
 
-        for(int i=0; i<names->size() ;i++){
+        for(size_t i=0; i<names->size() ;i++){
             DBG(cerr <<  "   name["<< i << "]: "<< (*names)[i] << endl);
         }
     }
